+++ src/main/kotlin/no/nav/syfo/dialogmotekandidat/database/VarselOutboxRecipientDao.kt
@@ -7,13 +7,11 @@ import org.springframework.jdbc.core.RowMapper
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate
 import org.springframework.stereotype.Repository
-import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 import java.time.LocalDateTime
 import java.util.UUID
 import javax.inject.Inject
 
-@Service
 @Transactional
 @Repository
 class VarselOutboxRecipientDao @Inject constructor(
