+++ src/main/kotlin/no/nav/syfo/dialogmotekandidat/scheduler/KandidatVarselResolver.kt
@@ -1,0 +1,137 @@
+package no.nav.syfo.dialogmotekandidat.scheduler
+
+import no.nav.syfo.consumer.narmesteleder.NarmesteLederRelasjonDTO
+import no.nav.syfo.consumer.narmesteleder.NarmesteLederService
+import no.nav.syfo.dialogmote.database.DialogmoteDAO
+import no.nav.syfo.dialogmotekandidat.database.RecipientSpec
+import no.nav.syfo.dialogmotekandidat.database.VarselOutboxEntry
+import no.nav.syfo.dialogmotekandidat.kafka.KafkaDialogmotekandidatEndring
+import no.nav.syfo.motebehov.MotebehovService
+import no.nav.syfo.motebehov.motebehovstatus.MotebehovStatusHelper
+import no.nav.syfo.oppfolgingstilfelle.OppfolgingstilfelleService
+import no.nav.syfo.varsel.esyfovarsel.domain.*
+import org.slf4j.LoggerFactory
+import org.springframework.stereotype.Component
+import java.time.LocalDate
+import javax.inject.Inject
+
+@Component
+class KandidatVarselResolver @Inject constructor(
+    private val narmesteLederService: NarmesteLederService,
+    private val oppfolgingstilfelleService: OppfolgingstilfelleService,
+    private val dialogmoteDAO: DialogmoteDAO,
+    private val motebehovService: MotebehovService,
+    private val motebehovStatusHelper: MotebehovStatusHelper,
+) {
+    fun resolve(entry: VarselOutboxEntry, endring: KafkaDialogmotekandidatEndring): List<RecipientSpec> {
+        val ansattFnr = endring.personIdentNumber
+        val narmesteLedere = narmesteLederService.getAllNarmesteLederRelations(ansattFnr) ?: emptyList()
+
+        return if (endring.kandidat) {
+            resolveForSend(entry, ansattFnr, narmesteLedere)
+        } else {
+            resolveForFerdigstill(ansattFnr, narmesteLedere)
+        }
+    }
+
+    private fun resolveForSend(
+        entry: VarselOutboxEntry,
+        ansattFnr: String,
+        narmesteLedere: List<NarmesteLederRelasjonDTO>,
+    ): List<RecipientSpec> {
+        val oppfolgingstilfelle = oppfolgingstilfelleService.getActiveOppfolgingstilfelleForArbeidstaker(ansattFnr)
+
+        val isDialogmoteAlleredePlanlagt = dialogmoteDAO.getAktiveDialogmoterEtterDato(
+            ansattFnr,
+            oppfolgingstilfelle?.fom ?: LocalDate.now(),
+        ).isNotEmpty()
+
+        if (isDialogmoteAlleredePlanlagt) {
+            log.info("Oppretter ingen mottakere for ${entry.uuid} — dialogmøte er allerede planlagt")
+            return emptyList()
+        }
+
+        val recipients = mutableListOf<RecipientSpec>()
+
+        val isSvarBehovAvailableForAT = motebehovStatusHelper.isSvarBehovVarselAvailable(
+            motebehovService.hentMotebehovListeForOgOpprettetAvArbeidstaker(ansattFnr),
+            oppfolgingstilfelle,
+        )
+        if (isSvarBehovAvailableForAT) {
+            recipients += RecipientSpec(
+                mottakerFnr = ansattFnr,
+                hendelse = ArbeidstakerHendelse(
+                    type = HendelseType.SM_DIALOGMOTE_SVAR_MOTEBEHOV,
+                    ferdigstill = false,
+                    data = null,
+                    arbeidstakerFnr = ansattFnr,
+                    orgnummer = null,
+                ),
+            )
+        }
+
+        narmesteLedere.forEach { nl ->
+            val oppfolgingstilfelleForLeder = oppfolgingstilfelleService.getActiveOppfolgingstilfelleForArbeidsgiver(
+                ansattFnr,
+                nl.virksomhetsnummer,
+            )
+            val isSvarBehovAvailableForNL = motebehovStatusHelper.isSvarBehovVarselAvailable(
+                motebehovService.hentMotebehovListeForArbeidstakerOpprettetAvLeder(ansattFnr, false, nl.virksomhetsnummer),
+                oppfolgingstilfelleForLeder,
+            )
+            if (isSvarBehovAvailableForNL) {
+                recipients += RecipientSpec(
+                    mottakerFnr = nl.narmesteLederPersonIdentNumber,
+                    hendelse = NarmesteLederHendelse(
+                        type = HendelseType.NL_DIALOGMOTE_SVAR_MOTEBEHOV,
+                        ferdigstill = false,
+                        data = null,
+                        narmesteLederFnr = nl.narmesteLederPersonIdentNumber,
+                        arbeidstakerFnr = ansattFnr,
+                        orgnummer = nl.virksomhetsnummer,
+                    ),
+                )
+            }
+        }
+
+        return recipients
+    }
+
+    private fun resolveForFerdigstill(
+        ansattFnr: String,
+        narmesteLedere: List<NarmesteLederRelasjonDTO>,
+    ): List<RecipientSpec> {
+        val recipients = mutableListOf<RecipientSpec>()
+
+        recipients += RecipientSpec(
+            mottakerFnr = ansattFnr,
+            hendelse = ArbeidstakerHendelse(
+                type = HendelseType.SM_DIALOGMOTE_SVAR_MOTEBEHOV,
+                ferdigstill = true,
+                data = null,
+                arbeidstakerFnr = ansattFnr,
+                orgnummer = null,
+            ),
+        )
+
+        narmesteLedere.forEach { nl ->
+            recipients += RecipientSpec(
+                mottakerFnr = nl.narmesteLederPersonIdentNumber,
+                hendelse = NarmesteLederHendelse(
+                    type = HendelseType.NL_DIALOGMOTE_SVAR_MOTEBEHOV,
+                    ferdigstill = true,
+                    data = null,
+                    narmesteLederFnr = nl.narmesteLederPersonIdentNumber,
+                    arbeidstakerFnr = ansattFnr,
+                    orgnummer = nl.virksomhetsnummer,
+                ),
+            )
+        }
+
+        return recipients
+    }
+
+    companion object {
+        private val log = LoggerFactory.getLogger(KandidatVarselResolver::class.java)
+    }
+}
